Move dir_to_zipfile to dirtools
